# By leanid (24) and Leanid Chaika (2)
# Via leanid
* commit 'd4bde393d89112de1b05af4f847b6dae39301017 [formerly 7baabc89700204ba69028a34c882bd1300652bbd]': (26 commits)
  DF-8619 fix arrow buttons and review comments
  DF-8619 fix QE array kyes
  DF-8619 fix resource editor move and drag mouse buttons
  DF-8619 on mouse buttons up or down send windows pos not delta
  DF-8619 fix mouse buttons for win10
  DF-8619 fix megasolution build with input update
  DF-8619 fix testbed compile
  DF-8619 move inline function out class desc
  DF-8619 fix order of keyboard keys, add all new to end of enum
  DF-8619 minor revert to uppercase names and other minor changes
  DF-8619 revert UPPER_CASE enums back
  DF-8619 fix review max commet
  DF-8619 continue fixing ios compliling
  DF-8619 fix ios compile
  DF-8619 fix compile error
  DF-8619 fix review comments
  DF-8619 fix some reviews
  DF-8619 add lshift rshilft support to mac os tested on testbed
  DF-8619 minor fix
  DF-8619 fix win10 mouse move
  ...


Former-commit-id: f70fde7c9f266b9ee8d90ba5a2f79d2a49f7ff22

+++ Projects/TestBed/Classes/Tests/MultilineTest.cpp
@@ -127,9 +127,9 @@ void MultilineTest::LoadResources()
     AddControl(textField2);
     AddControl(textFieldMulti);
 
-    const uint32 Y_OFFSET = 500;
-    const uint32 CONTROL_LENGHT = 400;
-    const uint32 CONTROL_HEIGTH = 70;
+    const float32 Y_OFFSET = 500;
+    const float32 CONTROL_LENGHT = 400;
+    const float32 CONTROL_HEIGTH = 70;
 
     BaseScreen::LoadResources();
